requirements.txt
fastapi==0.104.0
pydantic==2.4.2
uvicorn==0.23.2
asyncpg==0.28.0
bcrypt==4.0.1
passlib==1.7.4
PyJWT==2.8.0
